pre_classifier/requirements.txt
open_clip_torch
torch
torchvision
opencv-python
pillow
numpy
pydantic
tqdm
python-dotenv
requests
pytest
